Don't implement TryInto<String>
The actual use cases need to ensure that the Term or TypedTerm is a binary or list, so use binary_to_string and list_to_string, which check the type too before conversion to string.
