+++ lumen_web/src/document/get_element_by_id_2.rs
@@ -1,4 +1,3 @@
-use std::convert::TryInto;
 use std::sync::Arc;
 
 use liblumen_alloc::erts::exception;
@@ -8,6 +7,8 @@ use liblumen_alloc::erts::process::code::{self, result_from_exception};
 use liblumen_alloc::erts::process::Process;
 use liblumen_alloc::erts::term::{Atom, Term};
 use liblumen_alloc::erts::ModuleFunctionArity;
+
+use lumen_runtime::binary_to_string::binary_to_string;
 
 use crate::document::document_from_term;
 use crate::option_to_ok_tuple_or_error;
@@ -67,7 +68,7 @@ fn module_function_arity() -> Arc<ModuleFunctionArity> {
 
 pub fn native(process: &Process, document: Term, id: Term) -> exception::Result {
     let document_document = document_from_term(document)?;
-    let id_string: String = id.try_into()?;
+    let id_string: String = binary_to_string(id)?;
 
     option_to_ok_tuple_or_error(process, document_document.get_element_by_id(&id_string))
         .map_err(|error| error.into())
